Move layer data into LayerPass for cleaner access
CURA-1601
